+++ example/example.dart
@@ -4,9 +4,9 @@ import 'package:neoansi/neoansi.dart';
 
 void main() {
   AnsiSink.from(stdout)
-    ..setForegroundColor(Ansi1BitColors.red)
+    ..setForegroundColor1(Ansi1BitColor.red)
     ..write('Hello ')
-    ..setForegroundColor(Ansi1BitColors.green)
+    ..setForegroundColor1(Ansi1BitColor.green)
     ..setUnderlined()
     ..write('World')
     ..resetStyles()