+++ example/progress_indicator.dart
@@ -3,12 +3,7 @@ import 'dart:io';
 import 'package:neoansi/neoansi.dart';
 
 void main() {
-  final console = AnsiSink.from(stdout);
-
-  // Make it easier to read the output.
-  console
-    ..clearScreen()
-    ..writeln('Loading...');
+  final console = AnsiSink.from(stdout)..writeln('Loading...');
 
   for (var i = 0; i < 100; i++) {
     sleep(const Duration(milliseconds: 10));