+++ lib/neoansi.dart
@@ -1,4 +1,5 @@
 import 'package:meta/meta.dart';
+import 'package:neocolor/neocolor.dart';
 
 part 'src/1_bit_colors.dart';
 part 'src/8_bit_colors.dart';